requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pandas
numpy
scikit-learn
requests
mlflow
dvc
python-dotenv
matplotlib

